refactor: extract label dialogue logic into service module (#332)
Move updateLabelDialogueHandler business logic (~400 lines) from
labels.routes.ts into a new service module labels/dialogue.ts.

Fixes:
- C1 (TOCTOU): all initial data fetching moved inside the transaction
  to eliminate race between read and lock.
- C2: move labelWordCounts cleanup from deleteLabelHandler into
  labels/crud.ts as cleanupLabelWordCounts().
- CL2: remove stale comment about UpdateLabelDialogueBody migration.
- CL3: add ValidationError handling to updateLabelDialogueHandler.

Oracle review: GO (no must-fix, all should-fix addressed).
